cleaning up `XunitLogger` instances
