+++ src/Akka.Hosting.Tests/ExtensionsSpecs.cs
@@ -8,8 +8,7 @@ using System;
 using System.Threading.Tasks;
 using Akka.Actor;
 using Akka.Configuration;
-using Akka.Event;
-using Akka.TestKit.Xunit2.Internals;
+using Akka.Hosting.TestKit.Internals;
 using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
