+++ src/Directory.Build.props
@@ -2,8 +2,9 @@
     <PropertyGroup>
         <Copyright>Copyright © 2013-2023 Akka.NET Team</Copyright>
         <Authors>Akka.NET Team</Authors>
-        <VersionPrefix>1.5.19</VersionPrefix>
-        <PackageReleaseNotes>• [Update Akka.NET to 1.5.19](https://github.com/akkadotnet/akka.net/releases/tag/1.5.19)</PackageReleaseNotes>
+        <VersionPrefix>1.5.26</VersionPrefix>
+        <PackageReleaseNotes>• [Update Akka.NET to 1.5.26](https://github.com/akkadotnet/akka.net/releases/tag/1.5.26)
+• Added scopes to all Microsoft.Extensions.Logging output%2C for better semantic logging support.</PackageReleaseNotes>
         <PackageIcon>akkalogo.png</PackageIcon>
         <PackageProjectUrl>
             https://github.com/akkadotnet/Akka.Hosting
